I think I fixed the indexing issue for collisions in the device kernel (might still have to fix cpu). I am attempting to return collision counter back to int.

+++ include/cuda_dsmc/collisions.hpp
@@ -56,7 +56,7 @@ void collide_atoms(int num_atoms,
                    int *cell_num_atoms,
                    int *cell_cumulative_num_atoms,
                    double *collision_remainder,
-                   double *collision_count);
+                   int *collision_count);
 
 void collide_atoms(int num_atoms,
                    int num_cells,
@@ -71,7 +71,7 @@ void collide_atoms(int num_atoms,
                    int *cell_num_atoms,
                    int *cell_cumulative_num_atoms,
                    double *collision_remainder,
-                   double *collision_count);
+                   int *collision_count);
 
 void index_atoms(int num_atoms,
                  double3 *pos,
@@ -108,7 +108,7 @@ void collide(int num_cells,
              int *cell_cumulative_num_atoms,
              double dt,
              curandState *state,
-             double *collision_count,
+             int *collision_count,
              double *collision_remainder,
              double  *sig_vr_max,
              double3 *vel);
@@ -119,7 +119,7 @@ void collide(int num_cells,
              int *cell_cumulative_num_atoms,
              double dt,
              pcg32_random_t *state,
-             double *collision_count,
+             int *collision_count,
              double *collision_remainder,
              double  *sig_vr_max,
              double3 *vel);
